+++ src/config/mapper.php
@@ -12,7 +12,9 @@ return [
          */
         'case_style' => 'snake_case',
     ],
-    'map_namespace' => '\App\Libraries\MapperV2\Maps',
+
+    'map_namespace' => '\App\Maps',
+
     /**
      * Map mapping dto and entity
      */